refactor: 핵심 로직을 도메인 패키지로 이동하라
