+++ src/main/kotlin/lotto/domain/LottoNumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 fun interface LottoNumberGenerator {
     fun generate(): Set<Int>
