+++ src/test/kotlin/lotto/LottoStoreTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
+import lotto.domain.LottoStore
 
 class LottoStoreTest : BehaviorSpec({
     val lottoStore = LottoStore()
