+++ src/test/kotlin/lotto/LottoTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
+import lotto.domain.Lotto
 
 class LottoTest : BehaviorSpec({
     Given("유효한 로또 번호로 생성") {
